add support for jinja templates
also rename template folder to fit

jinja templates have extension .jinja and I went with them to allow me to write
some more complex templates without having to do *everything* before the
template, as a result of the lack of basic logic in mustache. the template
system has not changed in terms of backwards compatability and the interface is
the same, however now it allows for mustache and jinja templates to be used
together in the same page. YAY